Refactor: layout를 common으로 이동, category 동적 라우트 재구성, components 파일 PascalCase 리네임
