--- src/app/(community)/community/page.tsx
@@ -1,5 +1,0 @@
-import { CommunityShell } from "@/app/(community)/community/components/shell/communityShell";
-
-export default function Page() {
-  return <CommunityShell category="all" />;
-}
