+++ src/components/common/layout/MainNav.tsx
@@ -4,10 +4,10 @@ import { usePathname } from "next/navigation";
 export default function MainNav() {
   const pathname = usePathname();
   const isPostActive = pathname === "/mypage/posts";
-  const isCutActive = pathname === "/community";
-  const isFreeActive = pathname === "/community/free";
-  const isQnaActive = pathname === "/community/qna";
-  const isInfoActive = pathname === "/community/info";
+  const isCutActive = pathname === "/community/category/all";
+  const isFreeActive = pathname === "/community/category/free";
+  const isQnaActive = pathname === "/community/category/qna";
+  const isInfoActive = pathname === "/community/category/info";
 
   return (
     <div
@@ -16,28 +16,28 @@ export default function MainNav() {
     >
       <div className="flex h-full w-[424px] items-center justify-between">
         <Link
-          href="/community"
+          href="/community/category/all"
           aria-current={isCutActive ? "page" : undefined}
           className="cursor-pointer text-gray-400 hover:font-bold hover:text-orange-400 aria-[current=page]:font-semibold aria-[current=page]:text-orange-300"
         >
           전체
         </Link>
         <Link
-          href="/community/free"
+          href="/community/category/free"
           aria-current={isFreeActive ? "page" : undefined}
           className="cursor-pointer text-gray-400 hover:font-bold hover:text-orange-400 aria-[current=page]:font-semibold aria-[current=page]:text-orange-300"
         >
           자유게시판
         </Link>
         <Link
-          href="/community/qna"
+          href="/community/category/qna"
           aria-current={isQnaActive ? "page" : undefined}
           className="cursor-pointer text-gray-400 hover:font-bold hover:text-orange-400 aria-[current=page]:font-semibold aria-[current=page]:text-orange-300"
         >
           질문게시판
         </Link>
         <Link
-          href="/community/info"
+          href="/community/category/info"
           aria-current={isInfoActive ? "page" : undefined}
           className="cursor-pointer text-gray-400 hover:font-bold hover:text-orange-400 aria-[current=page]:font-semibold aria-[current=page]:text-orange-300"
         >
