--- src/types/test.tsx
@@ -1,3 +1,0 @@
-export default function Test() {
-  return <div>로그인 페이지</div>;
-}
